requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pytest==7.4.3
pytest-django==4.7.0
psycopg2-binary==2.9.7
requests==2.31.0
